requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
async-timeout==4.0.2
attrs==22.2.0
blinker==1.5
blobfile==2.0.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.1
dataclasses-json==0.5.7
decorator==5.1.1
entrypoints==0.4
filelock==3.9.0
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
gpt-index==0.4.4
greenlet==2.0.2
huggingface-hub==0.12.0
idna==3.4
importlib-metadata==6.0.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
langchain==0.0.88
lxml==4.9.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.24.2
openai==0.26.5
packaging==23.0
pandas==1.5.3
pdfminer.six==20221105
pdfplumber==0.8.0
Pillow==9.4.0
protobuf==3.20.3
pyarrow==11.0.0
pycparser==2.21
pycryptodomex==3.17
pydantic==1.10.5
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
rich==13.3.1
semver==2.13.0
six==1.16.0
smmap==5.0.0
SQLAlchemy==1.4.46
streamlit==1.18.1
tenacity==8.1.0
tiktoken==0.2.0
tokenizers==0.13.2
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
transformers==4.26.1
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
validators==0.20.0
Wand==0.6.11
watchdog==2.2.1
yarl==1.8.2
zipp==3.13.0
unstructured
pinecone-client
